requirements.txt
requests>=2.8.1
vcrpy
